refactor: reorganize project folders
