--- daemon/grpc/main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"victord/daemon/grpc/server"
-)
-
-func main() {
-	server.StartGRPCServer()
-}
