--- daemon/grpc/proto/victord.proto
@@ -1,32 +1,0 @@
-syntax = "proto3";
-
-package pb;
-option go_package = "git@github.com/victor-base/victord/daemon/grpc/pb";
-
-
-message CreateIndexRequest {
-  uint32 index_type = 1;
-  uint32 method = 2;
-  uint32 dims = 3;
-  string index_name = 4;
-}
-
-message CreateIndexResponse {
-  string status = 1;
-  string message = 2;
-  repeated CreateIndexResult results = 3;
-
-}
-
-message CreateIndexResult {
-  string index_name = 1;
-  string id = 2;
-  uint32 dims = 3;
-  string index_type = 4;
-  string method = 5;
-}
-
-
-service VictorService {
-  rpc CreateIndex (CreateIndexRequest) returns (CreateIndexResponse);
-}
