+++ daemon/internal/entity/index/types.go
@@ -1,0 +1,9 @@
+package index
+
+type CreateIndexResult struct {
+	IndexName string `json:"index_name"`
+	ID        string `json:"id"`
+	Dims      uint16 `json:"dims"`
+	IndexType int    `json:"index_type"`
+	Method    int    `json:"method"`
+}
